start getting patch2self into workflows
