+++ qsiprep/cli/run.py
@@ -197,6 +197,11 @@ def get_parser():
         help='window size in voxels for ``dwidenoise``. Must be odd (default: 5). '
              'If 0, ``dwidwenoise`` will not be run')
     g_conf.add_argument(
+        '--denoise-method', '--denoise_method',
+        action='store',
+        choices=['dwidenoise', 'patch2self'],
+        help='Image-based denoising method. Either dwidenoise (MRtrix) or patch2self (DIPY)')
+    g_conf.add_argument(
         '--unringing-method', '--unringing_method',
         action='store',
         choices=['none', 'mrdegibbs'],
