+++ qsiprep/workflows/dwi/merge.py
@@ -16,6 +16,7 @@ from nipype.interfaces import utility as niu
 from .util import _get_wf_name
 from .qc import init_modelfree_qc_wf
 from ...interfaces import ConformDwi, DerivativesDataSink
+from ...interfaces.dipy import Patch2Self
 from ...interfaces.mrtrix import DWIDenoise, DWIBiasCorrect, MRDeGibbs
 from ...interfaces.gradients import ExtractB0s
 from ...interfaces.nilearn import MaskEPI, Merge
@@ -30,6 +31,7 @@ def init_merge_and_denoise_wf(raw_dwi_files,
                               dwi_denoise_window,
                               unringing_method,
                               dwi_no_biascorr,
+                              denoise_method,
                               no_b0_harmonization,
                               denoise_before_combining,
                               orientation,
@@ -49,6 +51,7 @@ def init_merge_and_denoise_wf(raw_dwi_files,
         wf = init_merge_and_dwnoise_wf(['/path/to/dwi/sub-1_dwi.nii.gz'],
                                        source_file='/data/sub-1/dwi/sub-1_dwi.nii.gz',
                                        dwi_denoise_window=7,
+                                       denoise_method='patch2self',
                                        b0_threshold=100,
                                        unringing_method='mrdegibbs',
                                        dwi_no_biascorr=False,
@@ -63,8 +66,10 @@ def init_merge_and_denoise_wf(raw_dwi_files,
         b0_threshold : int
             Maximum b value for an image to be considered a b=0
         dwi_denoise_window : int
-            window size in voxels for ``dwidenoise``. Must be odd. If 0,
-            ``dwidwenoise`` will not be run
+            window size in voxels for image-based denoising. Must be odd. If 0, '
+            'denoising will not be run'
+        denoise_method : str
+            Either 'dwidenoise', 'patch2self' or 'none'
         unringing_method : str
             algorithm to use for removing Gibbs ringing. Options: none, mrdegibbs
         dwi_no_biascorr : bool
@@ -138,6 +143,7 @@ def init_merge_and_denoise_wf(raw_dwi_files,
             wf_name = _get_wf_name(fname).replace('preproc', 'denoise')
             denoising_wfs.append(
                 init_dwi_denoising_wf(dwi_denoise_window=dwi_denoise_window,
+                                      denoise_method=denoise_method,
                                       unringing_method=unringing_method,
                                       dwi_no_biascorr=dwi_no_biascorr,
                                       no_b0_harmonization=no_b0_harmonization,
@@ -213,6 +219,7 @@ def init_merge_and_denoise_wf(raw_dwi_files,
     hstack_confounds = pe.Node(StackConfounds(axis=1), name='hstack_confounds')
     denoising_wf = init_dwi_denoising_wf(
         dwi_denoise_window=dwi_denoise_window,
+        denoise_method=denoise_method,
         unringing_method=unringing_method,
         dwi_no_biascorr=dwi_no_biascorr,
         no_b0_harmonization=no_b0_harmonization,
@@ -240,6 +247,7 @@ def init_merge_and_denoise_wf(raw_dwi_files,
 
 
 def init_dwi_denoising_wf(dwi_denoise_window,
+                          denoise_method,
                           unringing_method,
                           dwi_no_biascorr,
                           no_b0_harmonization,
@@ -274,7 +282,7 @@ def init_dwi_denoising_wf(dwi_denoise_window,
             ('bvec_file', 'bvec_file')])])
 
     # Which steps to apply?
-    do_denoise = dwi_denoise_window > 0
+    do_denoise = denoise_method in ('patch2self', 'dwidenoise')
     do_unringing = unringing_method in ('mrdegibbs', 'dipy')
     do_biascorr = not dwi_no_biascorr
     harmonize_b0s = not no_b0_harmonization
@@ -284,11 +292,15 @@ def init_dwi_denoising_wf(dwi_denoise_window,
 
     # Add the steps
     step_num = 1  # Merge inputs start at 1
-    if dwi_denoise_window > 0:
-        denoiser = pe.Node(
-            DWIDenoise(extent=(dwi_denoise_window, dwi_denoise_window,
-                               dwi_denoise_window)),
-            name='denoiser')
+    if do_denoise:
+        if denoise_method == 'dwidenoise':
+            denoiser = pe.Node(
+                DWIDenoise(extent=(dwi_denoise_window, dwi_denoise_window,
+                                dwi_denoise_window)),
+                name='denoiser')
+        else:
+            denoiser = pe.Node(Patch2Self(), name='denoiser')
+
         ds_report_denoising = pe.Node(
             DerivativesDataSink(suffix=name + '_denoising',
                                 source_file=source_file),
