+++ qsiprep/workflows/dwi/pre_hmc.py
@@ -28,6 +28,7 @@ def init_dwi_pre_hmc_wf(scan_groups,
                         b0_threshold,
                         preprocess_rpe_series,
                         dwi_denoise_window,
+                        denoise_method,
                         unringing_method,
                         dwi_no_biascorr,
                         no_b0_harmonization,
@@ -56,6 +57,7 @@ def init_dwi_pre_hmc_wf(scan_groups,
                                   b0_threshold=100,
                                   preprocess_rpe_series=False,
                                   dwi_denoise_window=7,
+                                  denoise_method='dwidenoise',
                                   unringing_method='mrdegibbs',
                                   dwi_no_biascorr=False,
                                   no_b0_harmonization=False,
@@ -123,6 +125,7 @@ def init_dwi_pre_hmc_wf(scan_groups,
                                                dwi_denoise_window=dwi_denoise_window,
                                                unringing_method=unringing_method,
                                                dwi_no_biascorr=dwi_no_biascorr,
+                                               denoise_method=denoise_method,
                                                no_b0_harmonization=no_b0_harmonization,
                                                denoise_before_combining=denoise_before_combining,
                                                orientation=orientation,
@@ -136,6 +139,7 @@ def init_dwi_pre_hmc_wf(scan_groups,
         merge_minus = init_merge_and_denoise_wf(raw_dwi_files=minus_files,
                                                 b0_threshold=b0_threshold,
                                                 dwi_denoise_window=dwi_denoise_window,
+                                                denoise_method=denoise_method,
                                                 unringing_method=unringing_method,
                                                 dwi_no_biascorr=dwi_no_biascorr,
                                                 no_b0_harmonization=no_b0_harmonization,
@@ -242,6 +246,7 @@ def init_dwi_pre_hmc_wf(scan_groups,
         raw_dwi_files=dwi_series,
         b0_threshold=b0_threshold,
         dwi_denoise_window=dwi_denoise_window,
+        denoise_method=denoise_method,
         unringing_method=unringing_method,
         dwi_no_biascorr=dwi_no_biascorr,
         no_b0_harmonization=no_b0_harmonization,
